# apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Mesh gateway
  version: 0.1.0
  summary: Trading fees become hourly AI inference credits, spent through an OpenAI-compatible gateway, served by Macs and frontier providers, and sold on a marketplace when unused.
  description: |
    Mesh is a token whose trading fees buy AI inference for the people holding it. A 1.5 % fee on every
    $MESH trade is swept once an hour; half goes to the network treasury and half is split, pro rata, across
    every wallet holding at least 1,000 tokens (time-weighted over the hour) as credits denominated in US
    dollars. A second engine, a share of the margin on paid requests and marketplace fees, pays into the same
    hourly pool (`GET /stats → usageShareEnabled`).
    Holders spend credits through this gateway with their own API keys: open models on Macs in the network at
    a flat price per million tokens, frontier models through zero-data-retention providers at list plus a 6 %
    markup that covers what the upstream charges Mesh (`GET /v1/models` carries both prices). Credits nobody
    will use are sold on the credit marketplace (`/market/*`, docs/MARKETPLACE.md), and anyone can buy credits
    from Mesh at face value with a prepaid balance (`/me/credits/buy`). A holder's first sign-in receives
    starter credits, which can be spent but not sold.

    What stands behind a credit, and what limits it:

    * **Reserve.** The holder share of every sweep is swapped to a stablecoin on chain and kept in a
...

...

...

...
servers:
...
security: []
paths:
...
  /auth/nonce:
    post:
      tags: [Auth]
      summary: Issue a sign-in message
      description: |
        Returns a SIWE / SIWS-style message for the wallet to sign. The nonce is single-use and expires in 5 minutes.
        Sign `message` exactly as returned, then call `POST /auth/verify`. Rate limited per IP (20/min by default).
      operationId: authNonce
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet]
              properties:
                wallet: { type: string, maxLength: 128, description: Base58 (Solana) or 0x address (EVM). }
            example: { wallet: 9xQeKf2sVbT7mRw3Lp8nHJq4cYzA6dEuGk1oXiSNHn4k }
      responses:
        '200':
          description: Message to sign.
          content:
            application/json:
              schema:
                type: object
                required: [wallet, nonce, domain, issuedAt, expiresAt, expiresInSec, message]
                properties:
                  wallet: { type: string }
                  nonce: { type: string }
                  domain: { type: string }
                  issuedAt: { type: string, format: date-time }
                  expiresAt: { type: string, format: date-time }
                  expiresInSec: { type: integer }
                  message: { type: string, description: The exact text the wallet must sign. }
              example:
                wallet: 9xQeKf2sVbT7mRw3Lp8nHJq4cYzA6dEuGk1oXiSNHn4k
                nonce: 7f3a2c91e4b8d0f6
                domain: app.example.com
                issuedAt: '2026-10-03T09:00:00.000Z'
                expiresAt: '2026-10-03T09:05:00.000Z'
                expiresInSec: 300
                message: "app.example.com wants you to sign in with your wallet:\n9xQeKf2sVbT7mRw3Lp8nHJq4cYzA6dEuGk1oXiSNHn4k\n\nSign in to Mesh.\n\nURI: https://app.example.com\nNonce: 7f3a2c91e4b8d0f6\nIssued At: 2026-10-03T09:00:00.000Z\nExpiration Time: 2026-10-03T09:05:00.000Z"
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/verify:
    post:
      tags: [Auth]
      summary: Exchange a signature for a session
      description: Verifies the ed25519 (Solana) or EIP-191 (EVM) signature over the issued message and returns a 7-day session JWT.
      operationId: authVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet, signature]
              properties:
                wallet: { type: string }
                signature: { type: string, description: Base58 (Solana) or 0x hex (EVM) signature over `message`. }
                chain: { type: string, enum: [solana, evm], description: Defaults to the gateway's configured chain. }
                nonce: { type: string, description: Optional; defaults to the newest live nonce for the wallet. }
                message: { type: string, maxLength: 2000, description: Optional echo of the signed message for precise error messages. }
                invite: { type: string, maxLength: 64, description: 'Beta: invite code for a wallet that has never been admitted (`403 invite_required` says when one is needed). Consumed once; the wallet stays admitted.' }
            example:
              wallet: 9xQeKf2sVbT7mRw3Lp8nHJq4cYzA6dEuGk1oXiSNHn4k
              signature: 3xT1…Qk9b
              chain: solana
              nonce: 7f3a2c91e4b8d0f6
      responses:
        '200':
          description: Session token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
              example: { token: eyJhbGciOiJIUzI1NiJ9…, wallet: 9xQeKf2sVbT7mRw3Lp8nHJq4cYzA6dEuGk1oXiSNHn4k, chain: solana, expiresIn: 7d, expiresInSec: 604800 }
        '400':
          description: '`nonce_missing` (expired, used or unknown), `domain_mismatch`, `wallet_mismatch`, `message_mismatch` or `bad_request`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: nonce_missing, message: 'nonce missing, expired (5 min) or already used; request a new one' }
        '401':
          description: '`bad_signature`: the signature does not match the issued message.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 'Public beta: `invite_required` (the signature is fine but this wallet has not been admitted; retry with `invite`, or join `POST /waitlist`) or `invite_invalid` (`reason`: `invalid` | `exhausted`).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: invite_required, message: 'Mesh is in beta: this wallet needs an invite code to sign in. Join the waitlist or enter your code.', statusCode: 403, beta: { enabled: true, label: Beta, inviteRequired: true } }
        '429': { $ref: '#/components/responses/RateLimited' }
  /waitlist:
    post:
      tags: [Auth]
      summary: Join the beta waitlist
      description: Public, rate-limited (5/min per IP). Takes a wallet or an e-mail; idempotent per identifier. `position` is 1-based among entries not yet invited (0 once invited). `404` when the beta is not running.
      operationId: waitlistJoin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                wallet: { type: string, minLength: 8, maxLength: 128 }
                email: { type: string, format: email }
            example: { email: you@example.com }
      responses:
        '200':
          description: Listed.
          content:
            application/json:
              example: { ok: true, position: 184, alreadyListed: false, beta: { enabled: true, label: Beta, inviteRequired: true } }
        '400':
          description: Neither wallet nor e-mail, or an invalid e-mail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: The beta waitlist is closed.
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /nodes/register/challenge:
    post:
      tags: [Nodes]
      summary: Challenge to sign before registering
      description: Step 1 of signed registration. The reward wallet signs `message`; single-use, 5 minutes. Rate limited per IP (`NODE_REGISTER_RATE_LIMIT`, 10/hour, shared with unsigned registrations; the per-IP backstop applies too).
      operationId: nodeChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet]
              properties:
                wallet: { type: string }
                nodeId: { type: string, pattern: '^[A-Za-z0-9_.:-]+$', description: Optional stable id to bind into the message. }
            example: { wallet: 9xQeKf2sVbT7mRw3Lp8nHJq4cYzA6dEuGk1oXiSNHn4k }
      responses:
        '200':
          description: Challenge.
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallet: { type: string }
                  nonce: { type: string }
                  nodeId: { type: [string, 'null'] }
                  domain: { type: string }
                  expiresAt: { type: string, format: date-time }
                  message: { type: string }
                  requireSignature: { type: boolean }
                  hint: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /nodes/register:
    post:
      tags: [Nodes]
      summary: Register a node
      description: |
        Called by `mesh-node`. Either `{ linkCode }` (browser flow) or `{ wallet, nonce, signature }` (signed challenge).
        Returns the node token once; the gateway stores only its hash. Re-registering an existing `nodeId` requires its token.
        `maxParallel` is how many jobs the node runs at once (Ollama `OLLAMA_NUM_PARALLEL`); capped by the gateway's
        `routing.maxParallelPerNode` (4), default 1; the effective value is echoed back.
        Rate limits: `NODE_REGISTER_RATE_LIMIT` (10/hour) is counted **per reward wallet** once the link code or signature
        has proved ownership (so one operator can set up many Macs behind one NAT), per IP for unsigned registrations and
        for `/challenge`; a per-IP backstop (`NODE_REGISTER_IP_RATE_LIMIT`, default max(60, 6 × that)) applies to every call.
      operationId: nodeRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                linkCode: { type: string }
                wallet: { type: string }
                nonce: { type: string }
                signature: { type: string }
                chain: { type: string, enum: [solana, evm] }
                nodeId: { type: string }
                chip: { type: string, example: M2 Pro }
                ramGb: { type: number, example: 32 }
                models: { type: array, items: { type: string }, example: ['llama3.1:8b'] }
                agentVersion: { type: string, example: 0.1.0 }
                maxParallel: { type: integer, minimum: 1, maximum: 64, default: 1, description: Concurrent jobs the node can run; capped by routing.maxParallelPerNode. }
            example: { linkCode: K7Q2M9XD, chip: M2 Pro, ramGb: 32, models: ['llama3.1:8b'], agentVersion: 0.1.0, maxParallel: 2 }
      responses:
        '200':
          description: Registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodeId: { type: string }
                  nodeToken: { type: string, description: 'Shown once.' }
                  wallet: { type: string }
                  registered: { type: boolean }
                  walletVerified: { type: boolean }
                  linked: { type: boolean }
                  heartbeatEverySec: { type: integer }
                  offlineAfterSec: { type: integer }
                  pollMaxWaitMs: { type: integer }
              example: { nodeId: node_7kd2pq9f, nodeToken: mesh_node_Qm3…, wallet: 9xQeKf2sVbT7mRw3Lp8nHJq4cYzA6dEuGk1oXiSNHn4k, registered: true, walletVerified: true, linked: true, heartbeatEverySec: 20, offlineAfterSec: 60, pollMaxWaitMs: 25000 }
        '400':
          description: '`link_code_invalid`, `link_code_used`, `link_code_expired`, `nonce_missing` or `bad_request`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: '`signature_required` or `bad_signature`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: '`node_exists`: re-registering this `nodeId` needs its node token.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 7-day session from `POST /auth/verify`.
    apiKey:
      type: http
      scheme: bearer
      description: '`mesh_sk_…` from `POST /keys`. Works wherever an OpenAI key works.'
    nodeToken:
      type: http
      scheme: bearer
      description: '`mesh_node_…` from `POST /nodes/register`.'
...
  responses:
    BadRequest:
      description: Validation failed. `issues` lists the fields.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: bad_request, issues: [{ path: [wallet], message: Required }] }
...
    RateLimited:
      description: Too many requests. `retry-after` says when to try again.
      headers:
        retry-after: { schema: { type: integer } }
        x-ratelimit-limit: { schema: { type: integer } }
        x-ratelimit-remaining: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: rate_limited, message: 'Too many auth requests; retry in a minute.', statusCode: 429 }
...
  schemas:
...
    Error:
      type: object
      description: Error shape outside `/v1`.
      required: [error]
      properties:
        error: { type: string, description: Machine-readable code. }
        message: { type: string }
        statusCode: { type: integer }
        requestId: { type: string }
        issues: { type: array, items: { type: object } }
...
    Session:
      type: object
      required: [token, wallet, chain, expiresIn, expiresInSec]
      properties:
        token: { type: string, description: 'JWT. Send as `Authorization: Bearer <token>`.' }
        wallet: { type: string }
        chain: { type: string, enum: [solana, evm] }
        expiresIn: { type: string, example: 7d }
        expiresInSec: { type: integer, example: 604800 }
        starter:
          type: [object, 'null']
          description: 'Set on the first-ever sign-in of a wallet when starter credits are on (config `starterCredits`, docs/SWITCHING.md): the USD just credited and the resulting balance. `null` on every later sign-in, when the wallet does not hold the minimum, when the programme is paused or its wallet cap is reached. Starter credit can be spent on requests but not sold.'
          properties:
            amountUsd: { type: number, example: 2 }
            balanceUsd: { type: number, example: 2 }
...